requirements.txt
# langchain-openai
langchain_core
python-dotenv
streamlit
langchain_community
langserve
langchain-ollama